Move some opal test to analysis package and remove unused resources
