--- analyzer/javacg-opal/src/test/resources/maven-resolver-test-pom/test-pom-profiles.txt
@@ -1,36 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <groupId>org.example</groupId>
-    <artifactId>profiletest</artifactId>
-    <packaging>jar</packaging>
-    <version>1.0-SNAPSHOT</version>
-    <name>Maven Quick Start Archetype</name>
-    <url>http://maven.apache.org</url>
-    <properties>
-        <dep.version>1.7.30</dep.version>
-    </properties>
-    <dependencies></dependencies>
-    <profiles>
-        <profile>
-            <id>Profile</id>
-            <dependencies>
-                <dependency>
-                    <groupId>org.slf4j</groupId>
-                    <artifactId>slf4j-simple</artifactId>
-                    <version>${dep.version}</version>
-                </dependency>
-                <dependency>
-                    <groupId>org.dom4j</groupId>
-                    <artifactId>dom4j</artifactId>
-                </dependency>
-                <dependency>
-                    <groupId>info.picocli</groupId>
-                    <artifactId>picocli</artifactId>
-                    <version>1.0.0</version>
-                </dependency>
-            </dependencies>
-        </profile>
-    </profiles>
-</project>
